requirements.txt
PySide6==6.7.2
pyautogui==0.9.54
mss==9.0.1
Pillow==10.4.0
pytesseract==0.3.13
GitPython==3.1.43
pyperclip==1.8.2
